3d/requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
numpy-stl>=3.0.0
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0
scikit-learn>=1.1.0
